updated code for authentication and User services
