+++ package.json
@@ -9,8 +9,10 @@
   "author": "bhaskar,shivam",
   "license": "ISC",
   "dependencies": {
+    "axios": "^1.6.2",
     "bcrypt": "^5.1.1",
     "cookie-parser": "^1.4.6",
+    "cors": "^2.8.5",
     "express": "^4.18.2",
     "express-session": "^1.17.3",
     "mongoose": "^8.0.3",
